+++ src/Plugin.php
@@ -67,7 +67,7 @@ class Plugin extends BasePlugin
     /**
      * @inheritdoc
      */
-    public string $schemaVersion = '5.0.0';
+    public string $schemaVersion = '5.0.0.1';
 
     /**
      * @inheritdoc